Rename mockpb dir to mockbm
