+++ src/mockbm/network.py
