+++ src/tests/mockbm/pybitmessage/singleton.py
